fireship-analyzer/pyproject.toml
[project]
name = "fireship-analyzer"
version = "0.1.0"
description = "Analyze Fireship YouTube transcripts to extract content creation guidelines"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["analyzer"]
